libbpf: add support to dedup to generate and use minimal base BTF
When set, this option makes BTF deduplication generate a minimal base
BTF - the minimal base BTF needed to allow split BTF to refer to it -
and deduplication is carried out against that minimal base BTF.

If deduplication succeeds, references from split BTF will be to the
types in the _minimal_ base BTF.  The minimal base BTF is now the base
BTF associated with the "struct btf *" passed into deduplication, and
can be retrieved via btf__base_btf(btf).  Carrying a minimal base
BTF representation allows split BTF to be more resilient to base
BTF changes; if no minimal base BTF is used, and type ids change
in the base BTF, the split BTF is useless, since there is no context
associated with the base BTF ids.  However,  if we  carry the
minimal base BTF as the view of base BTF at deduplication time,
compatibility checks between minimal base BTF and current base BTF
can be carried out.  If compatibility checks succeed, the split
BTF can be used, with updated base BTF id references referring
to the current base BTF.

This form of deduplication is useful for cases such as when a kernel
module is compiled less frequently than the underlying vmlinux from
which it gets base BTF.  By carrying the minimal base BTF representation
of base BTF at the time of BTF encoding, we are in a position to
assess compatibility at the time of use.

Generation of minimal base BTF is reasonably straightforward - the
hypot_map is used to map from original base BTF to updated BTF
type ids.  All split BTF type references that point at base BTF
type ids in the deduplication map are remapped to point at minimal
base BTF, and the split ids themselves are remapped to start just
after the minimal base BTF ids.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>